+++ proto/gen_svr/error_svr.cs
@@ -39,7 +39,7 @@ namespace abelkhan
 
         public void error_code(error err_code){
             var _argv_b2890040_4e10_3967_a5ca_fac1ca306870 = new ArrayList();
-            _argv_b2890040_4e10_3967_a5ca_fac1ca306870.Add(err_code);
+            _argv_b2890040_4e10_3967_a5ca_fac1ca306870.Add((int)err_code);
             hub.hub._gates.call_client(client_uuid_f657c200_2a33_308f_b478_82cef68fbca2, "error_code_ntf_error_code", _argv_b2890040_4e10_3967_a5ca_fac1ca306870);
         }
