+++ proto/gen_svr/playercallmatch_svr.cs
@@ -50,7 +50,7 @@ namespace abelkhan
 
         public void player_join_match(playground _playground, player_inline_info player_info){
             var _argv_53045b75_544f_3315_95b9_4f6ed684c17d = new ArrayList();
-            _argv_53045b75_544f_3315_95b9_4f6ed684c17d.Add(_playground);
+            _argv_53045b75_544f_3315_95b9_4f6ed684c17d.Add((int)_playground);
             _argv_53045b75_544f_3315_95b9_4f6ed684c17d.Add(player_inline_info.player_inline_info_to_protcol(player_info));
             hub.hub._hubs.call_hub(hub_name_f08f93cd_bfea_3bf2_ae83_42be38c1f420, "player_match_player_join_match", _argv_53045b75_544f_3315_95b9_4f6ed684c17d);
         }
